Java/Java Fundamentals/03.Java OOP Advanced - July 2018/Preparation/ArrPeeGee_IvoHristov/src/main/java/app/models/participants/Warrior.java
package app.models.participants;

import app.constants.Config;

public class Warrior extends AbstractHero {

    public Warrior() {
        super(Config.WARRIOR_BASE_STRENGTH, Config.WARRIOR_BASE_DEXTERITY, Config.WARRIOR_BASE_INTELLIGENCE);
    }

    public Warrior(String name) {
        this();
        this.setName(name);
    }

    @Override
    public double getDamage() {
        return (this.getStrength() * Config.WARRIOR_DAMAGE_MULTIPLIER) + this.getDexterity();
    }

    @Override
    public void takeDamage(double damage) {
        super.takeDamage(damage);
        super.activateSpeciality();
    }
}
